requirements.txt
click
gitpython
pyyaml
ipdb
jinja2
boto3
